python_code/api/requirements.txt
pandas
python-dotenv==1.0.1
openai==1.61.1
runpod==1.7.7
pinecone
